Retry locking if wait fails within timeout. Reset ttl after wait.
This implies:
- Resetting the TTL requires knowing the key before we enter the
pipeline, which means we cant have it generated by redis during the
aquireLock anymore. So that needs to be generated client-side now.
- the acquireScript needs to extend the TTL if it's already owned, since
we may be re-trying after a failed WAIT. At which point we may or may
not have the lock in the current master already. And we'll need to be
sure that the ttl will survive the next WAIT. This implicitly gives
aquireLock the ability to extend a lock.